refactor: organize code for article & TOC
- encapsulated articles' logic and change file name(posts -> article)
- move the TOC generator to a separated util file
